+++ apps/api/src/app/inbox/usecases/session/session.usecase.ts
@@ -75,7 +75,7 @@ import { SessionCommand } from './session.command';
 
 const ALLOWED_ORIGINS_REGEX = new RegExp(process.env.FRONT_BASE_URL || '');
 const KEYLESS_RETENTION_TIME_IN_HOURS = parseInt(process.env.KEYLESS_RETENTION_TIME_IN_HOURS || '', 10) || 24;
-const MAX_NOTIFICATIONS_COUNT = 99;
+const MAX_NOTIFICATIONS_COUNT = 100;
 
 @Injectable()
 export class Session {
@@ -121,12 +121,6 @@ export class Session {
       throw new BadRequestException('Please provide a valid application identifier');
     }
 
-    const contextKeys = await this.resolveContexts(
-      environment._id,
-      environment._organizationId,
-      command.requestData.context
-    );
-
     const inAppIntegration = await this.selectIntegration.execute(
       SelectIntegrationCommand.create({
         environmentId: environment._id,
@@ -156,6 +150,12 @@ export class Session {
         });
       }
     }
+
+    const contextKeys = await this.resolveContexts(
+      environment._id,
+      environment._organizationId,
+      command.requestData.context
+    );
 
     const subscriberEntity = await this.createSubscriber.execute(
       CreateOrUpdateSubscriberCommand.create({
@@ -420,7 +420,11 @@ export class Session {
       return [];
     }
 
-    const contexts = await this.contextRepository.upsertContextsFromPayload(environmentId, organizationId, context);
+    const contexts = await this.contextRepository.findOrCreateContextsFromPayload(
+      environmentId,
+      organizationId,
+      context
+    );
 
     return contexts.map((context) => context.key);
   }