+++ apps/api/src/app/integrations/usecases/generate-chat-oath-url/generate-chat-oauth-url.command.ts
@@ -1,0 +1,18 @@
+import { ChatProviderIdEnum, ResourceKey } from '@novu/shared';
+import { IsEnum, IsNotEmpty, IsString } from 'class-validator';
+import { EnvironmentCommand } from '../../../shared/commands/project.command';
+import { IsResourceKey } from '../../../shared/validators/resource-key.validator';
+
+export class GenerateChatOauthUrlCommand extends EnvironmentCommand {
+  @IsNotEmpty()
+  @IsString()
+  readonly integrationIdentifier: string;
+
+  @IsNotEmpty()
+  @IsEnum(ChatProviderIdEnum)
+  readonly providerId: ChatProviderIdEnum;
+
+  @IsNotEmpty()
+  @IsResourceKey()
+  readonly resource: ResourceKey;
+}