+++ libs/internal-sdk/src/funcs/environmentsGetTags.ts
@@ -119,7 +119,7 @@ async function $do(
     options: client._options,
     baseURL: options?.serverURL ?? client._baseURL ?? '',
     operationID: 'EnvironmentsController_getEnvironmentTags',
-    oAuth2Scopes: [],
+    oAuth2Scopes: null,
 
     resolvedSecurity: requestSecurity,