+++ libs/internal-sdk/src/funcs/subscribersPropertiesUpdateOnlineFlag.ts
@@ -124,7 +124,7 @@ async function $do(
     options: client._options,
     baseURL: options?.serverURL ?? client._baseURL ?? '',
     operationID: 'SubscribersV1Controller_updateSubscriberOnlineFlag',
-    oAuth2Scopes: [],
+    oAuth2Scopes: null,
 
     resolvedSecurity: requestSecurity,